+++ packages/entity_store/lib/src/store.dart
@@ -3,6 +3,7 @@ library store;
 import 'dart:async';
 
 import 'package:collection/collection.dart';
+import 'package:equatable/equatable.dart';
 import 'package:flutter/foundation.dart';
 import 'package:state_notifier/state_notifier.dart';