+++ packages/entity_store_firestore/lib/entity_store_firestore.dart
@@ -5,3 +5,4 @@ export 'src/firestore_repository/firestore_entity_notifier.dart';
 export 'src/firestore_repository.dart';
 export 'src/in_memory_repository.dart';
 export 'src/repository_interface.dart';
+export 'src/local_storage_repository.dart';