+++ packages/entity_store_firestore/lib/src/firestore_repository.dart
@@ -2,7 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:collection/collection.dart';
 import 'package:entity_store/entity_store.dart';
 import 'package:entity_store_firestore/entity_store_firestore.dart';
-import 'package:skyreach_result/skyreach_result.dart';
+import 'package:type_result/type_result.dart';
 
 part 'firestore_repository/query.dart';
 part "firestore_repository/options.dart";