+++ packages/entity_store_firestore/lib/src/firestore_repository/firestore_entity_notifier.dart
@@ -2,7 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:entity_store/entity_store.dart';
 import 'package:entity_store_firestore/entity_store_firestore.dart';
 import 'package:meta/meta.dart';
-import 'package:skyreach_result/skyreach_result.dart';
+import 'package:type_result/type_result.dart';
 
 abstract class IFirestoreEntityNotifier<Id, E extends Entity<Id>> {
   @protected
@@ -286,22 +286,22 @@ extension _BoolX on bool {
 }
 
 extension _BoolOrNullX on bool? {
-  T ifMap<T>({
-    required T Function() ifTrue,
-    required T Function() ifFalse,
-    required T Function() ifNull,
-  }) {
-    if (this == null) {
-      return ifNull();
-    } else if (this!) {
-      return ifTrue();
-    } else {
-      return ifFalse();
-    }
-  }
+  // T ifMap<T>({
+  //   required T Function() ifTrue,
+  //   required T Function() ifFalse,
+  //   required T Function() ifNull,
+  // }) {
+  //   if (this == null) {
+  //     return ifNull();
+  //   } else if (this!) {
+  //     return ifTrue();
+  //   } else {
+  //     return ifFalse();
+  //   }
+  // }
 
-  /// if null, return false
-  bool get orFalse => this ?? false;
+  // /// if null, return false
+  // bool get orFalse => this ?? false;
 
   /// if null, return true
   bool get orTrue => this ?? true;