+++ packages/entity_store_firestore/lib/src/in_memory_repository.dart
@@ -1,6 +1,6 @@
 import 'package:entity_store/entity_store.dart';
 import 'package:entity_store_firestore/src/in_memory_repository/query.dart';
-import 'package:skyreach_result/skyreach_result.dart';
+import 'package:type_result/type_result.dart';
 
 import 'repository_interface.dart';