+++ packages/entity_store_firestore/lib/src/local_storage_repository.dart
@@ -1,0 +1,10 @@
+import 'package:collection/collection.dart';
+import 'package:entity_store/entity_store.dart';
+import 'package:type_result/type_result.dart';
+
+import 'local_storage_repository/local_storage_entity_hander.dart';
+import 'repository_interface.dart';
+
+part 'local_storage_repository/repository.dart';
+part 'local_storage_repository/query.dart';
+part 'local_storage_repository/local_storage_handler.dart';