+++ packages/entity_store_firestore/lib/src/local_storage_repository/local_storage_handler.dart
@@ -1,0 +1,8 @@
+part of '../local_storage_repository.dart';
+
+abstract class LocalStorageHandler {
+  Future<Result<void, Exception>> save(String key, String value);
+  Future<Result<String?, Exception>> load(String key);
+  Future<Result<void, Exception>> delete(String key);
+  Future<Result<void, Exception>> clear();
+}