+++ packages/entity_store_firestore/lib/src/repository_interface.dart
@@ -1,5 +1,5 @@
 import 'package:entity_store/entity_store.dart';
-import 'package:skyreach_result/skyreach_result.dart';
+import 'package:type_result/type_result.dart';
 
 part 'repository_interface/repository.dart';
 part 'repository_interface/query.dart';