+++ packages/entity_store_firestore/lib/src/repository_interface/repository.dart
@@ -21,7 +21,7 @@ abstract class IRepository<Id, E extends Entity<Id>> {
   });
 
   Future<Result<Id, Exception>> delete(
-    Id entity, {
+    Id id, {
     IDeleteOptions? options,
   });