+++ packages/entity_store/CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1-dev.1
 
 * init
+
+## 1.0.2
+
+* Changed EntityStore and EntityMap to inherit from Equatable
+
+## 2.0.0
+
+* Change the library of Result type