+++ packages/entity_store_firestore/CHANGELOG.md
@@ -1,3 +1,19 @@
 ## 0.0.1
 
 * TODO: Describe initial release.
+
+## 1.0.2
+
+* Changed the version of the entity_store package to 1.0.2
+
+## 2.0.0
+
+* Change the library of Result type
+
+## 2.1.0
+
+* Added LocalStorageRepository for storing Entity in local storage
+
+## 2.1.1
+
+* Fix import